backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pymupdf==1.24.3
sentence-transformers==2.7.0
faiss-cpu==1.8.0
rank-bm25==0.2.2
numpy==1.26.4
groq==0.5.0
torch==2.3.0
transformers==4.41.0